requirements.txt
requests==2.25.0
python-nmap==0.6.1
flask==1.1.2
gevent==20.9.0
pymongo==3.11.1
multiprocessing==2.6.2.1
apscheduler==3.6.3
func-timeout==4.3.5
pocsuite==2.0.8
argparse==1.4.0
basic==0.1
logging==0.4.9.6
flask-swagger==0.2.14
msgpack==1.0.0
qqwry==1.0
bs4==0.0.1
ipaddr==2.2.0